requirements.txt
streamlit

slack_bolt
slack_sdk
openai
pydantic
unstructured


langchain
langchain_community
langchain-core
langgraph
langchain_openai
langchain_google_community[drive]
langchain-groq
langchain-anthropic
langchain-xai

google-auth
google-auth-oauthlib
google-api-python-client
google-auth-httplib2
google-cloud-firestore

python-dateutil
markdown2

instructor>=0.4.0
altair
numpy
pandas
pydeck
streamlit
PyPDF2
pinecone-client
langchain
langchain-core 
langchain-community 
langchain-openai
openai
sendgrid
markdown
mistune
pydrive